// src/types/gen.ts
export interface components {
// ...
  responses: {
    /** There was an error */
    ErrorResponse: {
      content: {
        "application/json": {
          statusCode?: number;
          /** Specific description of the error */
          error?: string;
          /** What the error was */
          message?: string;
          /** Some extra information about the error */
          data?: { [key: string]: any };
        };
      };
    };
  };
}

export interface operations {
  /** Returns in descending order of most recently used */
  tokenGet: {
    responses: {
      /** OK */
      200: {
        content: {
          "application/json": components["schemas"]["RefreshToken"][];
        };
      };
      400: components["responses"]["ErrorResponse"];
      401: components["responses"]["ErrorResponse"];
      403: components["responses"]["ErrorResponse"];
      500: components["responses"]["ErrorResponse"];
    };
  };
  tokenPost: {
// ...
    requestBody: {
      content: {
        "application/json": components["schemas"]["AuthRequest"];
      };
    };
  };
  /** If the token is not specified, all tokens of the user are revoked */
  tokenDelete: {
    parameters: {
      query: {
        token?: string;
      };
    };
    responses: {
      /** Revoked */
      204: never;
      400: components["responses"]["ErrorResponse"];
      401: components["responses"]["ErrorResponse"];
      403: components["responses"]["ErrorResponse"];
      500: components["responses"]["ErrorResponse"];
    };
  };
// ...
  inviteLinksGet: {
    parameters: {
      path: {
        /** the invite link ID */
        id: string;
      };
    };
    responses: {
      /** OK */
      200: {
        content: {
          "application/json": components["schemas"]["InviteLink"];
        };
      };
      400: components["responses"]["ErrorResponse"];
      401: components["responses"]["ErrorResponse"];
      403: components["responses"]["ErrorResponse"];
      500: components["responses"]["ErrorResponse"];
    };
  };
// ...
}